% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parser.R
\docType{class}
\name{Parser}
\alias{Parser}
\title{Download api v2 parser}
\description{
Parse the JSON response from cube/report v2 download api and convert the data into a R data frame
}
\examples{
\dontrun{
# Create a parser object
p <- Parser$new(response)

# Parse the first request
p$parse(response)

# Return the data frame
p$to_dataframe()
}
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{response}}{JSON response from the cube/report v2 download api}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{response}}{JSON response from the cube/report v2 download api}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Parser$new()}}
\item \href{#method-parse}{\code{Parser$parse()}}
\item \href{#method-to_dataframe}{\code{Parser$to_dataframe()}}
\item \href{#method-clone}{\code{Parser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$new(response)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parse"></a>}}
\if{latex}{\out{\hypertarget{method-parse}{}}}
\subsection{Method \code{parse()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$parse(response)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-to_dataframe}{}}}
\subsection{Method \code{to_dataframe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$to_dataframe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
