% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-filter.R
\docType{class}
\name{Filter}
\alias{Filter}
\title{Pull MicroStrategy cubes (full or filtered)}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
Filter
}
\description{
Pass ids of selected objects (attributes, metrics and elements).
}
\section{Fields}{

\describe{
\item{\code{attributes}}{List of ids for selected attributes.}

\item{\code{metrics}}{List of ids for selected metrics.}

\item{\code{attr_elements}}{List of ids for selected attribute elements.}
}}

\keyword{datasets}
